+++ src/constants/Config.ts
@@ -1,0 +1,6 @@
+export const Config = {
+  root: "demo-time",
+  highlight: {
+    borderColor: "highlightBorderColor",
+  },
+};