+++ src/constants/index.ts
@@ -1,4 +1,5 @@
 export * from "./Command";
+export * from "./Config";
 export * from "./ContextKeys";
 export * from "./General";
 export * from "./StateKeys";